refs #3696 URL and Handle are now clickable in Resource Presentation
refs #3701
Changes in Resource Presentation applied. Texts are modified
Changes in Citation Page applied. Texts modified
Resources are reintegrated in class package because in Wicket "Resources need to place this in - or relative to - the same package as the component they are meant for".
Accordingly CMDINOdeIconProvider is reverted to previous code for accessing resources access.